embed tests moved
