+++ packages/ai/src/embed/embed.test.ts
@@ -3,8 +3,6 @@ import assert from 'node:assert';
 import { beforeEach, describe, expect, it, vi, vitest } from 'vitest';
 import * as logWarningsModule from '../logger/log-warnings';
 import { MockEmbeddingModelV4 } from '../test/mock-embedding-model-v4';
-import { MockTracer } from '../test/mock-tracer';
-import { OpenTelemetryIntegration } from '../telemetry/open-telemetry-integration';
 import { Embedding, EmbeddingModelUsage, Warning } from '../types';
 import { embed } from './embed';
 import type { EmbedOnStartEvent, EmbedOnFinishEvent } from './embed-events';
@@ -221,65 +219,6 @@ describe('logWarnings', () => {
       provider: 'mock-provider',
       model: 'mock-model-id',
     });
-  });
-});
-
-describe('telemetry', () => {
-  let tracer: MockTracer;
-
-  beforeEach(() => {
-    tracer = new MockTracer();
-  });
-
-  it('should not record any telemetry data when not explicitly enabled', async () => {
-    await embed({
-      model: new MockEmbeddingModelV4({
-        doEmbed: mockEmbed([testValue], [dummyEmbedding]),
-      }),
-      value: testValue,
-      experimental_telemetry: {
-        integrations: [new OpenTelemetryIntegration({ tracer })],
-      },
-    });
-
-    expect(tracer.jsonSpans).toMatchSnapshot();
-  });
-
-  it('should record telemetry data when enabled', async () => {
-    await embed({
-      model: new MockEmbeddingModelV4({
-        doEmbed: mockEmbed([testValue], [dummyEmbedding], { tokens: 10 }),
-      }),
-      value: testValue,
-      experimental_telemetry: {
-        isEnabled: true,
-        functionId: 'test-function-id',
-        metadata: {
-          test1: 'value1',
-          test2: false,
-        },
-        integrations: [new OpenTelemetryIntegration({ tracer })],
-      },
-    });
-
-    expect(tracer.jsonSpans).toMatchSnapshot();
-  });
-
-  it('should not record telemetry inputs / outputs when disabled', async () => {
-    await embed({
-      model: new MockEmbeddingModelV4({
-        doEmbed: mockEmbed([testValue], [dummyEmbedding], { tokens: 10 }),
-      }),
-      value: testValue,
-      experimental_telemetry: {
-        isEnabled: true,
-        recordInputs: false,
-        recordOutputs: false,
-        integrations: [new OpenTelemetryIntegration({ tracer })],
-      },
-    });
-
-    expect(tracer.jsonSpans).toMatchSnapshot();
   });
 });
